requirements.txt
playwright
google-generativeai
pyyaml
